+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This CHANGELOG details important changes made in each version of the
 - Allow users to specify a specific provider version.
 - Add the ability to deprecate resources and datasources.
 - Emit an appropriate user warning when Pulumi binary not found in Python setup.py.
+- Add support for suppressing differences between the desired and actual state of a resource via the `ignoreChanges` property.
 
 ## v0.18.3 (Released June 20, 2019)